+++ cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/repository/redis/AuthorizationRepositoryRedis.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * @author ziyao
  */
 @Repository
-public interface AuthorizationRepositoryRedis extends RedisValueRepository<Authorization, Long> {
+public interface AuthorizationRepositoryRedis extends RedisValueRepository<Authorization, Integer> {
 }
