+++ cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/service/security/AuthorizationCenter.java
@@ -12,13 +12,13 @@ public interface AuthorizationCenter {
     /**
      * 授权
      */
-    OAuth2AuthorizationCodeResponse authorize(Long appId, String state, String grantType);
+    OAuth2AuthorizationCodeResponse authorize(Integer appId, String state, String grantType);
 
     /**
      * 通过授权码或刷新token获取认证token
      *
      * @param authentication 授权对象
-     * @return {@link com.ziyao.security.oauth2.core.OAuth2AccessToken}
+     * @return {@link com.ziyao.ideal.security.oauth2.core.OAuth2AccessToken}
      */
     AccessTokenResponse token(Authentication authentication);
 }
